% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lump_lowfreq2.R
\name{fct_lump_lowfreq2}
\alias{fct_lump_lowfreq2}
\title{Lump together infrequent levels}
\usage{
fct_lump_lowfreq2(f, other_level = "Other")
}
\arguments{
\item{f}{A factor}

\item{other_level}{The name to use for the "Other" level.  Defaults to
\code{"Other"}.}
}
\value{
A reordered factor
}
\description{
This function is a variant of \code{fct_lump_infreq} from forcats, but instead of
lumping together levels such that the "Other" category is smaller than all
others, it lumps together levels such that the "Other" category is smaller
than the most frequent level
}
\examples{
x <- factor(rep(LETTERS[1:9], times = c(40, 10, 5, 27, 1, 1, 1, 1, 1)))
x \%>\% table()
x \%>\% fct_lump_lowfreq() \%>\% table()
# (Borrowed from forcats.)
}
